requirements.txt
# Core
langgraph
langchain
langchain-openai
pydantic-settings
pydantic
jinja2

# API
fastapi
uvicorn[standard]
python-dotenv
PyYAML

